DELETE exec-scpicmds (Qt) -> REPLACED BY exec-scpidmds-python
Explanation on why to remove this code:

The initial version for exec-scpitool was based on Qt and was only able
to execute static SCPi-command-files (one SCPI-command per line and no
evaluation of the responses).

Then the idea was to allow conditional loops and remain downwards
compatible. This was only possible by inventing commands like LOOP and
IF. These commands got extended even more and many ideas were even not
(completely) implemented.

The reason to stop this was that the user then has to learn this new
kind or programming language, it's a lot of work to extend and maintain
this code, it might not be welltested enough and it would make sense to
use an already established script language instead.

As an alternative, is was implemented in python and two modes were
introduced:
1) Execute plain SCPI-command-files.
2) Execute python files implementing a certain interface that gets then
   executed by the program. This allows to program the creation and
execution of SCPI-commands and at the same time use the full
functionality and behavior of the program. Super useful.

R.I.P. my dear :) Maybe you'll get resurrected some time for some other
purpose o.O

Signed-off-by: Anton Höß <a.hoess@zera.de>